+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueAppointment.kt
@@ -162,6 +162,25 @@ data class OverdueAppointment(
     ): PagingSource<Int, OverdueAppointment>
 
     @Query("""
+      $OVERDUE_APPOINTMENTS_QUERY 
+      WHERE
+        IFNULL(patientAssignedFacilityUuid, appt_facilityUuid) = :facilityUuid AND
+        (appt_scheduledDate > :scheduledAfter AND appt_scheduledDate < :scheduledBefore) AND
+        appt_status == "scheduled" AND
+        call_result_outcome = "agreed_to_visit"
+      GROUP BY appt_patientUuid
+      ORDER BY 
+        isAtHighRisk DESC, 
+        appt_scheduledDate DESC, 
+        appt_updatedAt ASC
+    """)
+    fun agreedToVisitOverdueAppointmentsInFacility(
+        facilityUuid: UUID,
+        scheduledAfter: LocalDate,
+        scheduledBefore: LocalDate
+    ): PagingSource<Int, OverdueAppointment>
+
+    @Query("""
       $OVERDUE_APPOINTMENTS_QUERY
       WHERE
         (