+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Bump Google play services auth to v20.3.0
 - Bump datadog gradle plugin to v1.5.1
 - Add query for getting pending appointments as a paging source
+- Add query for getting agreed to visit appointments as a paging source
 
 ### Fixes